+++ crate/cfg/src/clean_op_spec.rs
@@ -20,12 +20,12 @@ pub trait CleanOpSpec {
 
     /// Logical state of the managed item.
     ///
-    /// This is the type returned by the [`StateNowFnSpec`], and is used by
+    /// This is the type returned by the [`StateCurrentFnSpec`], and is used by
     /// [`EnsureOpSpec`] to determine if [`exec`] needs to be run.
     ///
     /// See [`FullSpec::StateLogical`] for more detail.
     ///
-    /// [`StateNowFnSpec`]: crate::FullSpec::StateNowFnSpec
+    /// [`StateCurrentFnSpec`]: crate::FullSpec::StateCurrentFnSpec
     /// [`EnsureOpSpec`]: crate::FullSpec::EnsureOpSpec
     /// [`exec`]: Self::exec
     /// [`FullSpec::StateLogical`]: crate::FullSpec::StateLogical
@@ -73,9 +73,10 @@ pub trait CleanOpSpec {
     /// # Parameters
     ///
     /// * `data`: Runtime data that the operation reads from, or writes to.
-    /// * `state`: State of the managed item, returned from [`StateNowFnSpec`].
+    /// * `state`: State of the managed item, returned from
+    ///   [`StateCurrentFnSpec`].
     ///
-    /// [`StateNowFnSpec`]: crate::FullSpec::StateNowFnSpec
+    /// [`StateCurrentFnSpec`]: crate::FullSpec::StateCurrentFnSpec
     /// [`StatePhysical`]: Self::StatePhysical
     async fn check(
         data: Self::Data<'_>,
@@ -98,12 +99,13 @@ pub trait CleanOpSpec {
     /// # Parameters
     ///
     /// * `data`: Runtime data that the operation reads from, or writes to.
-    /// * `state`: State of the managed item, returned from [`StateNowFnSpec`].
+    /// * `state`: State of the managed item, returned from
+    ///   [`StateCurrentFnSpec`].
     ///
     /// [`check`]: Self::check
     /// [`exec`]: Self::exec
     /// [`ExecRequired`]: crate::OpCheckStatus::ExecRequired
-    /// [`StateNowFnSpec`]: crate::FullSpec::StateNowFnSpec
+    /// [`StateCurrentFnSpec`]: crate::FullSpec::StateCurrentFnSpec
     async fn exec_dry(
         data: Self::Data<'_>,
         state: &State<Self::StateLogical, Self::StatePhysical>,
@@ -116,11 +118,12 @@ pub trait CleanOpSpec {
     /// # Parameters
     ///
     /// * `data`: Runtime data that the operation reads from, or writes to.
-    /// * `state`: State of the managed item, returned from [`StateNowFnSpec`].
+    /// * `state`: State of the managed item, returned from
+    ///   [`StateCurrentFnSpec`].
     ///
     /// [`check`]: Self::check
     /// [`ExecRequired`]: crate::OpCheckStatus::ExecRequired
-    /// [`StateNowFnSpec`]: crate::FullSpec::StateNowFnSpec
+    /// [`StateCurrentFnSpec`]: crate::FullSpec::StateCurrentFnSpec
     async fn exec(
         data: Self::Data<'_>,
         state: &State<Self::StateLogical, Self::StatePhysical>,