+++ crate/rt/src/commands.rs
@@ -1,5 +1,7 @@
-pub use self::{diff_cmd::DiffCmd, state_desired_cmd::StateDesiredCmd, state_now_cmd::StateNowCmd};
+pub use self::{
+    diff_cmd::DiffCmd, state_current_cmd::StateCurrentCmd, state_desired_cmd::StateDesiredCmd,
+};
 
 mod diff_cmd;
+mod state_current_cmd;
 mod state_desired_cmd;
-mod state_now_cmd;