+++ crate/rt/src/commands/state_current_cmd.rs
@@ -10,18 +10,19 @@ use peace_rt_model::FullSpecGraph;
 use crate::BUFFERED_FUTURES_MAX;
 
 #[derive(Debug)]
-pub struct StateNowCmd<E>(PhantomData<E>);
+pub struct StateCurrentCmd<E>(PhantomData<E>);
 
-impl<E> StateNowCmd<E>
+impl<E> StateCurrentCmd<E>
 where
     E: std::error::Error,
 {
-    /// Runs [`FullSpec`]`::`[`StateNowFnSpec`]`::`[`exec`] for each full spec.
+    /// Runs [`FullSpec`]`::`[`StateCurrentFnSpec`]`::`[`exec`] for each full
+    /// spec.
     ///
     /// At the end of this function, [`Resources`] will be populated with
     /// [`States`].
     ///
-    /// If any `StateNowFnSpec` needs to read the `State` from a previous
+    /// If any `StateCurrentFnSpec` needs to read the `State` from a previous
     /// `FullSpec`, the predecessor should insert a copy / clone of their state
     /// into `Resources`, and the successor should references it in their
     /// [`FnSpec::Data`].
@@ -29,7 +30,7 @@ where
     /// [`exec`]: peace_cfg::FnSpec::exec
     /// [`FnSpec::Data`]: peace_cfg::FnSpec::Data
     /// [`FullSpec`]: peace_cfg::FullSpec
-    /// [`StateNowFnSpec`]: peace_cfg::FullSpec::StateNowFnSpec
+    /// [`StateCurrentFnSpec`]: peace_cfg::FullSpec::StateCurrentFnSpec
     pub async fn exec(
         full_spec_graph: &FullSpecGraph<E>,
         resources: Resources<SetUp>,
@@ -39,14 +40,15 @@ where
         Ok(Resources::<WithStates>::from((resources, states)))
     }
 
-    /// Runs [`FullSpec`]`::`[`StateNowFnSpec`]`::`[`exec`] for each full spec.
+    /// Runs [`FullSpec`]`::`[`StateCurrentFnSpec`]`::`[`exec`] for each full
+    /// spec.
     ///
     /// Same as [`Self::exec`], but does not change the type state, and returns
     /// [`States`].
     ///
     /// [`exec`]: peace_cfg::FnSpec::exec
     /// [`FullSpec`]: peace_cfg::FullSpec
-    /// [`StateNowFnSpec`]: peace_cfg::FullSpec::StateNowFnSpec
+    /// [`StateCurrentFnSpec`]: peace_cfg::FullSpec::StateCurrentFnSpec
     pub(crate) async fn exec_internal(
         full_spec_graph: &FullSpecGraph<E>,
         resources: &Resources<SetUp>,
@@ -55,7 +57,7 @@ where
             .stream()
             .map(Result::Ok)
             .map_ok(|full_spec| async move {
-                let state = full_spec.state_now_fn_exec(resources).await?;
+                let state = full_spec.state_current_fn_exec(resources).await?;
                 Ok((full_spec.id(), state))
             })
             .try_buffer_unordered(BUFFERED_FUTURES_MAX)