+++ workspace_tests/src/rt.rs
@@ -1,3 +1,3 @@
 mod diff_cmd;
+mod state_current_cmd;
 mod state_desired_cmd;
-mod state_now_cmd;