+++ CHANGELOG.md
@@ -4,6 +4,6 @@
 
 ### Added
 
-* `StateNowCmd` to discover the current state.
+* `StateCurrentCmd` to discover the current state.
 * `StateDesiredCmd` to discover the desired state.
 * `DiffCmd` to compute the difference between the current and desired states.